[finishes #162445070] added validators and restructured code for challange 3
