+++ app/api/v1/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+api_bp = Blueprint("api", __name__)
