+++ app/api/v1/main/views.py
@@ -1,7 +1,8 @@
 # module import
 from flask_restful import Resource, reqparse
 
-from app.api.v1.models import Incident, User, incidents, Users
+from app.api.v1.models.incidents import Incident, incidents
+from app.api.v1.models.users import User
 from utils.validators import Validate
 
 
@@ -46,7 +47,7 @@ class PostIncidents(Resource):
         return {"incident": "created successfully"}, 201
 
 
-class Incidents(Resource):
+class AllIncidents(Resource):
 
     """ GET all incident """
 
@@ -115,71 +116,3 @@ class Get_incident_by_id(Resource):
             specific_incident.comment = comment
 
             return {"Incident": specific_incident.serializer()}, 200
-
-
-class Sign_in(Resource):
-    parser = reqparse.RequestParser()
-    
-    parser.add_argument('username', type=str, required=True)
-    parser.add_argument('password', type=str, required=True)
-
-    def post(self):
-        data = Sign_in.parser.parse_args()
-        username = data['username']
-        password = data['password']
-        user = User.get_user_by_username(username)
-        if user and user.password == password:
-            return {"Message": f"{user.username} created successfully"}, 201
-
-        return {"Message": "Username not found"}, 404
-
-
-class Sign_up(Resource):
-    parser = reqparse.RequestParser()
-
-    parser.add_argument('firstname', type=str, required=True)
-    parser.add_argument('lastname', type=str, required=True)
-    parser.add_argument('othernames', type=str, required=True)
-    parser.add_argument('email', type=str, required=True)
-    parser.add_argument('password', type=str, required=True)
-    parser.add_argument('phoneNumber', type=str, required=True)
-    parser.add_argument('username', type=str, required=True)
-    parser.add_argument('is_admin', type=bool, required=True)
-
-    def post(self):
-        data = Sign_up.parser.parse_args()
-        firstname = data['firstname']
-        lastname = data['lastname']
-        othernames = data['othernames']
-        email = data['email']
-        password = data['password']
-        phoneNumber = data['phoneNumber']
-        username = data['username']
-        is_admin = data['is_admin']
-
-        user = User(firstname, lastname, othernames, email, password,
-                    phoneNumber, username, bool(is_admin))
-        Users.append(user)
-        return {"Message": f"{user.username} created successfully"}, 201
-
-
-class All_users(Resource):
-    def get(self):
-        return {"Users": [user.serialize() for user in Users]}
-
-
-class Get_users_by_email(Resource):
-    def get(self, email):
-
-        user = User().get_user_by_email(email)
-        if user:
-            return {"User": user.serialize()}
-
-
-class Get_user_by_id(Resource):
-    def get(self, id):
-        user = User().get_user_by_id(id)
-        if not user:
-            return{"User does not exist"}
-        else:
-            return {"User": user.serialize()}
